added tests.h to mutualize tests utils.
