+++ tests/unit/bitarray_test.cpp
@@ -1,9 +1,10 @@
-#include <gtest/gtest.h>
+
+#include "tests.h"
 #include "bitarray.h"
 
 using namespace fuzzy_coco;
 
-auto sum = [](auto v) { return accumulate(v.begin(), v.end(), 0); };
+
 using namespace BitArrayUtils;
 TEST(bitarray, decode_number) {
   BitArray bits(8*100, false);
