+++ tests/unit/fuzzy_system_test.cpp
@@ -1,4 +1,4 @@
-#include <gtest/gtest.h>
+#include "tests.h"
 #include <algorithm>
 #include "file_utils.h"
 #include "fuzzy_system.h"
@@ -16,14 +16,6 @@ day3;24;NA;70
 day4;5;3;2
 )";
 
-template<typename T, class UnaryPred>
-bool all(const vector<T>& v, UnaryPred pred) { 
-  return std::all_of(v.cbegin(), v.cend(), pred);
-}
-
-template<typename T>
-bool all_zero(const vector<T>& v) { return all(v, [](T t) { return t == 0; });
-}
 
 // cf coco book pp26 fig 1.12
 string DB_BOOK_112 = R"(
