+++ tests/unit/named_list_test.cpp
@@ -1,4 +1,4 @@
-#include <gtest/gtest.h>
+#include "tests.h"
 #include "types.h"
 #include "named_list.h"
 #include "string_utils.h"
